doLogout re-defined in App.vue
I missed that the navbar logout button was moved back into App.vue in the previous commit.
